requirements.txt
streamlit==1.27.2
openai==0.28.1
llama-index==0.8.43
nltk==3.8.1
bs4==4.12.2
requests==2.31.0
python-dotenv==1.0.0